Warning fixes

+++ src/chips/opn_chip_base.cpp
@@ -29,7 +29,7 @@ int OPNChipBase::generate32(int32_t *output, size_t frames)
     enum { maxFramesAtOnce = 256 };
     int16_t temp[2 * maxFramesAtOnce];
     for(size_t left = frames; left > 0;) {
-        size_t count = (left < maxFramesAtOnce) ? left : maxFramesAtOnce;
+        size_t count = (left < (size_t)maxFramesAtOnce) ? left : (size_t)maxFramesAtOnce;
         int16_t *temp_it = temp;
         generate(temp, count);
         for(size_t i = 0; i < 2 * count; ++i)
@@ -44,7 +44,7 @@ int OPNChipBase::generateAndMix32(int32_t *output, size_t frames)
     enum { maxFramesAtOnce = 256 };
     int16_t temp[2 * maxFramesAtOnce];
     for(size_t left = frames; left > 0;) {
-        size_t count = (left < maxFramesAtOnce) ? left : maxFramesAtOnce;
+        size_t count = (left < (size_t)maxFramesAtOnce) ? left : (size_t)maxFramesAtOnce;
         int16_t *temp_it = temp;
         generate(temp, count);
         for(size_t i = 0; i < 2 * count; ++i)
